Author: Tigran Hovhannisyan
Changed the compilation process to use be used project name(root directory) to make directories for binary results and for dependencies. 

Changed the top makefile of the projects to use root directory name as project name.

Added cleaning of the bin and dep directories from tbuilder, by typing make clean in the top of tworkspace.