Author: Tigran Hovhannisyan
Added TODO in the top directory.
Added doc empty folder in the top.
Made changes in the exe.mk to link debugging information in case if the building is performed in the debug mode.
Made changes in the tutilities - added global declaration of functions.
Changed default mode to - coda-disable, debug-enable